[#198] refactor : useMutation api 로직 리팩토링
